+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+## v0.5.0 - 2024-01-14
+
+* Overhaul to attribute macro
+  * resources are now struct definitions
+  * each resource is explicitly extracted
+  * `#[cfg(...)]` along with other attribute macros work fine now
+  * add tests
+
 ## v0.4.0 - 2023-12-02
 
 * Rename generated struct to `AssignedResources` and make it pub